Use one entry point for `o-comments`
This was done to follow the Origami specs.
