+++ demos/src/auto-refresh.mustache
@@ -1,3 +1,6 @@
-<div data-o-component="o-comments-count" data-o-comments-count-config-article-id="{{articleId}}" data-o-comments-count-config-auto-refresh="true">
-Comments
+<div data-o-component="o-comments"
+  data-o-comments-count
+  data-o-comments-config-article-id="{{articleId}}"
+  data-o-comments-config-auto-refresh="true">
+  Comments
 </div>
