+++ demos/src/count.mustache
@@ -1,3 +1,5 @@
-<div data-o-component="o-comments-count" data-o-comments-count-config-article-id="{{articleId}}">
-Comments
+<div data-o-component="o-comments"
+  data-o-comments-count
+  data-o-comments-config-article-id="{{articleId}}">
+  Comments
 </div>
