+++ demos/src/custom-template.mustache
@@ -1,3 +1,6 @@
-<div data-o-component="o-comments-count" data-o-comments-count-config-article-id="{{articleId}}" data-o-comments-count-config-template="This article has {count} comment{plural}.">
-Comments
+<div data-o-component="o-comments"
+  data-o-comments-count
+  data-o-comments-config-article-id="{{articleId}}"
+  data-o-comments-config-template="This article has {count} comment{plural}.">
+  Comments
 </div>
